+++ src/merge_sort.h
@@ -5,6 +5,8 @@
 
 #include <omp.h>
 
+#include "utils.h"
+
 /**
  * Merge function
  * Merges two sorted arrays (left and right half)
@@ -14,11 +16,11 @@
  * @param mid Middle index of the subarray
  * @param right Right index of the subarray
  */
-void merge(std::vector<double> &arr, size_t left, size_t mid, size_t right);
+void merge(std::vector<decimal> &arr, size_t left, size_t mid, size_t right);
 
 /**
  * Modified merge sort function
  * Sorts the array and calculates the sum and sum of squares of the array elements
  * @param arr Array to be sorted
  */
-void merge_sort(std::vector<double> &arr);
+void merge_sort(std::vector<decimal> &arr);
